+++ server/src/models/play.rs
@@ -2,9 +2,9 @@
 #[derive(Debug)]
 pub struct Play {
     /// The id of the record.
-    pub id_play: usize,
+    pub id_play: i32,
     /// References the record in `tbl_player` that made the play.
-    pub id_player: usize,
+    pub id_player: i32,
     /// The letters that were added to the player's rack after
     /// the play.
     pub letters_added: String,
